+++ tests/testthat.R
@@ -3,3 +3,6 @@ library(hdWGCNA)
 
 #test_check("hdWGCNA")
 devtools::test()
+
+# remove files
+file.remove('tests/testthat/ConsensusTOM-block.1.rda')