+++ tests/testthat/test-SetDatExpr.R
@@ -3,22 +3,6 @@ test_that("SetDatExpr output is a data.frame", {
 
     data(test_seurat)
 
-    test_seurat <- SetupForWGCNA(
-        test_seurat,
-        wgcna_name = 'test',
-        features = rownames(test_seurat)
-    )
-
-    test_seurat <- MetacellsByGroups(
-        test_seurat,
-        group.by = c('cell_type'),
-        k = 5, 
-        max_shared = 5,
-        ident.group = 'cell_type',
-        target_metacells=50,
-        min_cells=10
-    )
-    test_seurat <- NormalizeMetacells(test_seurat, verbose=FALSE)
     test_seurat <- SetDatExpr(test_seurat, assay = 'RNA')
     datExpr <- GetDatExpr(test_seurat)
     expect_equal(class(datExpr), "data.frame")
@@ -30,22 +14,6 @@ test_that("features are columns", {
 
     data(test_seurat)
 
-    test_seurat <- SetupForWGCNA(
-        test_seurat,
-        wgcna_name = 'test',
-        features = rownames(test_seurat)
-    )
-
-    test_seurat <- MetacellsByGroups(
-        test_seurat,
-        group.by = c('cell_type'),
-        k = 5, 
-        max_shared = 5,
-        ident.group = 'cell_type',
-        target_metacells=50,
-        min_cells=10
-    )
-    test_seurat <- NormalizeMetacells(test_seurat, verbose=FALSE)
     test_seurat <- SetDatExpr(test_seurat, assay = 'RNA')
     datExpr <- GetDatExpr(test_seurat)
 
@@ -60,22 +28,6 @@ test_that("return_seurat flag works", {
 
     data(test_seurat)
 
-    test_seurat <- SetupForWGCNA(
-        test_seurat,
-        wgcna_name = 'test',
-        features = rownames(test_seurat)
-    )
-
-    test_seurat <- MetacellsByGroups(
-        test_seurat,
-        group.by = c('cell_type'),
-        k = 5, 
-        max_shared = 5,
-        ident.group = 'cell_type',
-        target_metacells=50,
-        min_cells=10
-    )
-    test_seurat <- NormalizeMetacells(test_seurat, verbose=FALSE)
     datExpr <- SetDatExpr(test_seurat, assay = 'RNA', return_seurat=FALSE)
     
     expect_equal(class(datExpr), "data.frame")
@@ -86,22 +38,6 @@ test_that("use_metacells flag works", {
 
     data(test_seurat)
 
-    test_seurat <- SetupForWGCNA(
-        test_seurat,
-        wgcna_name = 'test',
-        features = rownames(test_seurat)
-    )
-
-    test_seurat <- MetacellsByGroups(
-        test_seurat,
-        group.by = c('cell_type'),
-        k = 5, 
-        max_shared = 5,
-        ident.group = 'cell_type',
-        target_metacells=50,
-        min_cells=10
-    )
-    test_seurat <- NormalizeMetacells(test_seurat, verbose=FALSE)
     test_seurat <- SetDatExpr(test_seurat, assay = 'RNA', use_metacells=FALSE)
     datExpr <- GetDatExpr(test_seurat)